refact: - added data file and moved all constants into it - Moved constants into constants file - rewrote adding event handler
