+++ src/main.js
@@ -10,20 +10,19 @@ import { createFilmsCounterTemplate } from './view/films-counter-view';
 import { createFilmDetailsTemplate } from './view/film-details-view';
 import { generateFilm } from './mock/film';
 import { generateFilter } from './mock/filter';
+import { FILMS_LIST_COUNTER, FILMS_COUNTER_PER_STEP } from './const';
 
-const FILMS_LIST_COUNTER = 20;
-const FILMS_COUNTER_PER_STEP = 5;
 const header = document.querySelector('.header');
 const main = document.querySelector('.main');
 const footer = document.querySelector('.footer');
 const footerStatistics = footer.querySelector('.footer__statistics');
 
 const getAllFilms = () => {
-  const FILMS_LIST = [];
+  const filmsList = [];
   for (let i = 0; i < FILMS_LIST_COUNTER; i++) {
-    FILMS_LIST.push(generateFilm());
+    filmsList.push(generateFilm());
   }
-  return FILMS_LIST;
+  return filmsList;
 };
 
 const FILMS = getAllFilms();
@@ -47,7 +46,7 @@ for (let i = 0; i < Math.min(FILMS.length, FILMS_COUNTER_PER_STEP); i++) {
 const showFilmInfo = () => {
   const filmCards = document.querySelectorAll('.film-card');
   filmCards.forEach((filmCard, index) => {
-    const hangleOpen = (e) => {
+    const handleOpen = (e) => {
       const target = e.target;
 
       if (target.closest('.film-card__controls') || target.closest('.film-card__comments')) {
@@ -66,8 +65,7 @@ const showFilmInfo = () => {
       closePopupButton.addEventListener('click', handleClose);
     };
 
-    filmCard.removeEventListener('click', hangleOpen);
-    filmCard.addEventListener('click', hangleOpen);
+    filmCard.onclick = handleOpen;
   });
 };
 
@@ -94,19 +92,3 @@ if (FILMS.length > FILMS_COUNTER_PER_STEP) {
 renderTemplate(footerStatistics, createFilmsCounterTemplate(FILMS), RenderPosition.BEFOREEND);
 
 showFilmInfo();
-
-// console.log(generateFilter(FILMS));
-// console.log(FILMS);
-// const watcher = (mutationList) => {
-//   mutationList.forEach(mutation => {
-//     mutation.addedNodes.forEach(addedNode => {
-//       console.log(addedNode.classList.value);
-//     })
-//   });
-// };
-
-// const popupObserver = new MutationObserver(watcher);
-
-// popupObserver.observe(document.body, {
-//   childList: true,
-// });
