+++ src/mock/data.js
@@ -1,0 +1,106 @@
+export const names = [
+  'Venom 2',
+  'Not time to die',
+  'Thing',
+  'Duna',
+  'Eternals',
+];
+
+export const descriptions = [
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget.',
+  'Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.',
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra.',
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.',
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+];
+
+export const posters = [
+  'made-for-each-other.png',
+  'popeye-meets-sinbad.png',
+  'sagebrush-trail.jpg',
+  'santa-claus-conquers-the-martians.jpg',
+  'the-dance-of-life.jpg',
+];
+
+export const releases = [
+  {
+    date: '1929-05-11T00:00:00.000Z',
+    releaseCountry: 'USA',
+  },
+  {
+    date: '1933-06-14T00:00:00.000Z',
+    releaseCountry: 'United Kingdom',
+  },
+  {
+    date: '1955-03-21T00:00:00.000Z',
+    releaseCountry: 'Finland',
+  },
+  {
+    date: '1964-07-05T00:00:00.000Z',
+    releaseCountry: 'Russia',
+  },
+  {
+    date: '1936-10-01T00:00:00.000Z',
+    releaseCountry: 'China',
+  },
+];
+
+export const genre = [
+  ['Comedy', 'Action', 'Dramma'],
+  ['History', 'Action', 'Dramma'],
+  ['Dramma', 'Action', 'History'],
+  ['Action', 'Dramma', 'History'],
+  ['Action', 'Comedy', 'Dramma'],
+];
+
+export const directors = [
+  'Anthony Mann',
+  'Guy Ritchie',
+  'Robert Knepper',
+  'Rick Morty',
+  'Eliot Page',
+];
+
+export const actors = [
+  ['Morgan Freemen', 'Naomi Harris', 'Leonardo DiCaprio'],
+  ['Leonardo DiCaprio', 'Morgan Freemen', 'Kit Harrington'],
+  ['Tom Hardy', 'Leonardo DiCaprio', 'Morgan Freemen'],
+  ['Naomi Harris', 'Kit Harrington', 'Naomi Harris'],
+  ['Kit Harrington', 'Morgan Freemen', 'Leonardo DiCaprio'],
+];
+
+export const runtimes = [
+  120,
+  77,
+  159,
+  144,
+  100,
+];
+
+export const commentContent = [
+  'Good',
+  'a film that changed my life, a true masterpiece, post-credit scene was just amazing omg.',
+  'test comment',
+  'another test comment',
+];
+
+export const smiles = [
+  'smile',
+  'sleeping',
+  'puke',
+  'angry',
+];
+
+export const datesOfComment = [
+  '2019-05-11T16:13:48.554Z',
+  '2019-05-11T16:14:37.554Z',
+  '2019-05-11T16:15:21.554Z',
+  '2019-05-11T16:12:32.554Z',
+];
+
+export const authors = [
+  'Ilya O\'Reilly',
+  'Evelin Moss',
+  'Sergey Kuzmichev',
+  'Alexey Tsvetkov',
+];
